+++ babel/src/lib.rs
@@ -67,9 +67,8 @@ pub trait BabelPal {
     ) -> Result<()>;
     async fn is_data_drive_mounted(&self, data_directory_mount_point: &str) -> Result<bool>;
     async fn set_hostname(&self, hostname: &str) -> Result<()>;
-    async fn set_swap_file(&self, swap_size_mb: usize, swap_file_location: &str) -> Result<()>;
-    async fn is_swap_file_set(&self, swap_size_mb: usize, swap_file_location: &str)
-        -> Result<bool>;
+    async fn set_swap_file(&self, swap_size_mb: u64, swap_file_location: &str) -> Result<()>;
+    async fn is_swap_file_set(&self, swap_size_mb: u64, swap_file_location: &str) -> Result<bool>;
     async fn set_ram_disks(&self, ram_disks: Option<Vec<RamdiskConfiguration>>) -> Result<()>;
     async fn is_ram_disks_set(&self, ram_disks: Option<Vec<RamdiskConfiguration>>) -> Result<bool>;
 }
